Allow for configurable batch sizes for testing and change consistency test to use smaller batch sizes
This allows for a smaller "wide" consistency test

+++ pkg/util/chunking.go
@@ -5,15 +5,15 @@ import (
 )
 
 // ForEachChunk executes the given handler for each chunk of items in the slice.
-func ForEachChunk[T any](data []T, chunkSize uint64, handler func(items []T)) {
+func ForEachChunk[T any](data []T, chunkSize uint16, handler func(items []T)) {
 	if chunkSize == 0 {
 		logging.Warn().Int("invalid-chunk-size", int(chunkSize)).Msg("ForEachChunk got an invalid chunk size; defaulting to 1")
 		chunkSize = 1
 	}
 
-	dataLength := uint64(len(data))
+	dataLength := uint16(len(data))
 	chunkCount := (dataLength / chunkSize) + 1
-	for chunkIndex := uint64(0); chunkIndex < chunkCount; chunkIndex++ {
+	for chunkIndex := uint16(0); chunkIndex < chunkCount; chunkIndex++ {
 		chunkStart := chunkIndex * chunkSize
 		chunkEnd := (chunkIndex + 1) * chunkSize
 		if chunkEnd > dataLength {
